+++ spec/knossos/algorithm/binary_tree_spec.rb
@@ -1,0 +1,5 @@
+require 'knossos/algorithm/shared_examples_for_algorithm'
+
+RSpec.describe Knossos::Algorithm::BinaryTree do
+  it_behaves_like 'algorithm'
+end